Move net::parser tests to their own files
